refactor: update restaurant controller class
